documentation, fixing bugs and added additional respoonse types in the BasicCharInfoController

+++ backend/dndapp/src/main/java/com/scanakispersonalprojects/dndapp/controller/BasicCharInfoController.java
@@ -1,5 +1,6 @@
 package com.scanakispersonalprojects.dndapp.controller;
 
+import java.util.List;
 import java.util.UUID;
 import java.util.logging.Logger;
 
@@ -76,14 +77,20 @@ public class BasicCharInfoController {
      * @param uuid          unique idneitifer of the character
      * @param patch         The filed that are to be updated
      * @return              200 with an instance of the updated {@link CharacterBasicInfoView}
+     *                      401 if user not authorized
      *                      404 when the character does not exist
      *                      500 on any unexpected server error
      */
 
 
-   @PutMapping("/{uuid}")
-   public ResponseEntity<CharacterBasicInfoView> updateCharacterBasicView(@PathVariable UUID uuid , @RequestBody CharViewPatch patch) {
-            LOG.info("PUT /character/" + uuid);
+    @PutMapping("/{uuid}")
+    public ResponseEntity<CharacterBasicInfoView> updateCharacterBasicView(Authentication authentication, @PathVariable UUID uuid , @RequestBody CharViewPatch patch) {
+        LOG.info("PUT /character/" + uuid);
+        List<UUID> characters = userService.getUsersCharacters(authentication);
+        if(!characters.contains(uuid)) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             CharacterBasicInfoView charInfoView = charService.updateCharInfo(uuid, patch);
             if (charInfoView != null) {
@@ -96,13 +103,42 @@ public class BasicCharInfoController {
             LOG.severe(e::getMessage);
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
-   }
+    }
+
+   /**
+    * 
+    * Deletes a character and it's association with a User
+    *
+    * @param authentication     this endpoint requires that the user authenticating is the owner of the character
+    * @param uuid               the UUID of the character
+    * @return                   200 if deleted
+    *                           401 if user not authorized
+    *                           404 when the character not found
+    *                           500 on any unexpected server error
+    */
 
     @DeleteMapping("/{uuid}")
-    public boolean deleteCharacter(Authentication authentication ,@PathVariable UUID uuid) {
-        UUID userUuid = userService.getUsersUuid(authentication.getName());
+    public ResponseEntity<CharacterBasicInfoView> deleteCharacter(Authentication authentication ,@PathVariable UUID uuid) {
+        
+        List<UUID> characters = userService.getUsersCharacters(authentication);
+        if(!characters.contains(uuid)) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+ 
+        UUID userUuid = userService.getUsersUuid(authentication);
 
-        return charService.deleteCharacter(userUuid, uuid);
+        try {
+            boolean result = charService.deleteCharacter(userUuid, uuid);
+            if(result != false) {
+                return new ResponseEntity<>(HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
+         
     }
